examples/art-e/pyproject.toml
[project]
name = "art-e"
version = "0.1.0"
description = "Deep learning research for email search and processing"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=3.4.1",
    "diskcache>=5.6.3",
    "huggingface-hub>=0.29.3",
    "ipykernel>=6.29.5",
    "kaggle>=1.5.12",
    "langchain-core>=0.3.51",
    "mail-parser>=3.15.0",
    "matplotlib>=3.10.1",
    "pandas>=1.3.0",
    "panza>=0.1.0",
    "pip>=25.0.1",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "tabulate>=0.9.0",
    "tiktoken>=0.9.0",
    "tqdm>=4.62.0",
    "transformers>=4.50.3",
    "openpipe==4.49.0",
    "litellm>=1.65.0.post1",
    "ipywidgets>=8.1.6",
    "polars>=1.27.1",
    "openpipe-art",
    "langfuse>=2.60.5",
    "omegaconf>=2.4.0.dev3",
    "skypilot[runpod]>=0.9",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["art_e*"]

[tool.uv.sources]
art-e = { workspace = true }
openpipe-art = { path = "../../", editable = true }

[dependency-groups]
dev = ["art-e"]
